% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colleges}
\alias{colleges}
\title{colleges}
\format{A dataframe of 7,804 observations and 19 variables}
\source{
\url{https://collegescorecard.ed.gov/data/documentation/}
}
\usage{
colleges
}
\description{
Dataframe of all the colleges included in the College Scorecard, along with attributes such as geography,
religious affliation, predominant degree awarded, and more. Users are encouraged to use this dataset for
producing vectors of schools for use in other functions.
}
\examples{
data(colleges)
subset(colleges,colleges$HBCU==1)[c(2)]

}
\keyword{datasets}

