% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completionRate.R
\name{completionRate}
\alias{completionRate}
\title{Returns a bar chart for median debt levels for selected demographic groups by school. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard. Leave bygroup field blank if only interested
overall median.}
\usage{
completionRate(apiKey, dataset, schools, bygroup = "")
}
\arguments{
\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}

\item{bygroup}{Leave this blank to see completion rates overall, otherwise populate with 'race' to see completion by race}
}
\description{
Returns a bar chart for median debt levels for selected demographic groups by school. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard. Leave bygroup field blank if only interested
overall median.
}
\examples{
data(scorecard13)
completionRate(,scorecard13,c("University of Chicago","Northwestern University"),"race")
}

