% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDevNameToVarName.R
\name{convertDevNameToVarName}
\alias{convertDevNameToVarName}
\title{Returns the VARIABLE.NAME corresponding to the developer.friendly.name. When using the REST API, only the developer.friendly.name works when 
querying the data. However, when using the data included in the package, only the VARIABLE.NAME is available in the dataset. This
functions provides an easy way to get the developer.friendly.name when only the VARIABLE.NAME is known.}
\usage{
convertDevNameToVarName(devFriendlyName)
}
\arguments{
\item{devFriendlyName}{developer.friendly.name as per data dictionary}
}
\value{
character
}
\description{
Returns the VARIABLE.NAME corresponding to the developer.friendly.name. When using the REST API, only the developer.friendly.name works when 
querying the data. However, when using the data included in the package, only the VARIABLE.NAME is available in the dataset. This
functions provides an easy way to get the developer.friendly.name when only the VARIABLE.NAME is known.
}
\examples{
convertDevNameToVarName("8_yr_completion.low_income")
}

