% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costByIncome.R
\name{costByIncome}
\alias{costByIncome}
\title{Plots comparison of the average net price for different shcools based on student income level. The income levels are as follows:
Bracket 1: $0 - $30000 
Bracket 2: $30001 - $48000
Bracket 3: $48001 - $75000
Bracket 4: $75001 - $110000
Bracket 5: $110001-plus}
\usage{
costByIncome(apiKey, dataset, year = 2013, schoolNames, type = "public")
}
\arguments{
\item{apiKey}{If using API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{year}{Year of college scorecard data}

\item{schoolNames}{The names of schoold to be compared}

\item{type}{Type of institution. CUrrently the available options are: public, private}
}
\value{
ggplot2 object
}
\description{
Plots comparison of the average net price for different shcools based on student income level. The income levels are as follows:
Bracket 1: $0 - $30000 
Bracket 2: $30001 - $48000
Bracket 3: $48001 - $75000
Bracket 4: $75001 - $110000
Bracket 5: $110001-plus
}
\examples{
data(scorecard13)
costByIncome(dataset = scorecard13, year = 2013, schoolNames = 
c("University of Massachusetts-Lowell", "University of Arizona", "Alabama A & M University"))
costByIncome(dataset = scorecard13, year = 2013, schoolNames = c("Massachusetts Institute of 
Technology","Drake University", "Stanford University"), type = "private")
}

