% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataDict}
\alias{dataDict}
\title{Data Dictionary}
\format{A dataframe of 1,953 observations and 9 variables}
\source{
\url{https://collegescorecard.ed.gov/assets/CollegeScorecardDataDictionary-09-08-2015.csv}
}
\usage{
dataDict
}
\description{
Provides all variables and their definitions within the College Scorecard data.
}
\examples{
data(dataDict)
scorecardInfo(c("CITY","INSTNM","UNITID"))

}
\keyword{datasets}

