% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debtBoxplot.R
\name{debtBoxplot}
\alias{debtBoxplot}
\title{Returns a box plot of the 10th, 25th, 50th, 75th, and 90th percentiles of student debt by school. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard.}
\usage{
debtBoxplot(apiKey, dataset, schools)
}
\arguments{
\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}
}
\description{
Returns a box plot of the 10th, 25th, 50th, 75th, and 90th percentiles of student debt by school. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard.
}
\examples{
data(scorecard13)
debtBoxplot(,scorecard13,c("New York University","Cornell University"))
}

