% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femaleMaleEarnings.R
\name{femaleMaleEarnings}
\alias{femaleMaleEarnings}
\title{Output a plot using ggplot2 displaying earnings per school for female and male students for either 6 or 10 years 
after entry to a college/university. If specifying multiple years, please follow the format c(year1, year2, year3). 
When multiple years are specified, a plot will be displayed for each year. If schoolNames is not specified,
data for all schools will be returned. If used with the data provided with this pavkage, this API will produce no
results since no data is available for variables in the earnings group}
\usage{
femaleMaleEarnings(apiKey, dataset, year = 2013, schoolNames,
  yearsAfterCompletion = 6)
}
\arguments{
\item{apiKey}{If using API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{year}{Year of college scorecard data}

\item{schoolNames}{The names of schoold to be compared}

\item{yearsAfterCompletion}{Available values are 6 and 10}
}
\value{
ggplot2 object
}
\description{
Output a plot using ggplot2 displaying earnings per school for female and male students for either 6 or 10 years 
after entry to a college/university. If specifying multiple years, please follow the format c(year1, year2, year3). 
When multiple years are specified, a plot will be displayed for each year. If schoolNames is not specified,
data for all schools will be returned. If used with the data provided with this pavkage, this API will produce no
results since no data is available for variables in the earnings group
}
\examples{
\dontrun{femaleMaleEarnings(apiKey = ak, year = 2003, 
schoolNames = c("University of Massachusetts-Lowell", "Massachusetts Institute of Technology"))}
\dontrun{femaleMaleEarnings(apiKey = ak, year = c(2003, 2005, 2007, 2009, 2011), 
schoolNames = c("University of Massachusetts-Lowell", "Massachusetts Institute of Technology"))}
}

