% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllDataInCategory.R
\name{getAllDataInCategory}
\alias{getAllDataInCategory}
\title{Get all data from a specified category. The available categories are 
root, school, academics, admissions, student, cost, aid, repayment, completion, earnings}
\usage{
getAllDataInCategory(apiKey, categoryName, year, pattern = "",
  addParams = "id,school.name")
}
\arguments{
\item{apiKey}{Key used for interacting with the API}

\item{year}{The year(s) for data retrieval.}

\item{catgoryName}{The category of data for retrieval. Options include: academics, admission, aid, cost, earnings, repayment,
root, school, or student.}
}
\value{
character vector
}
\description{
Get all data from a specified category. The available categories are 
root, school, academics, admissions, student, cost, aid, repayment, completion, earnings
}
\examples{
costData <- getAllDataInCategory(categoryName = "cost", year = 2013)
schoolData <- getAllDataInCategory(categoryName = "school")
earningsData <- getAllDataInCategory(categoryName = "earnings", year = c(2010, 2013))
earningsData <- getAllDataInCategory(categoryName = "earnings", year = c(2010, 2013), pattern = "6_yrs_after_entry.mean", addParams = "school.state")
}

