% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIKey.R
\name{getAPIKey}
\alias{getAPIKey}
\title{Before calling this function an apiKey needs to be obtained from https://api.data.gov/signup"
and set using setAPIKey
This method loads the saved apiKey from key.rda}
\usage{
getAPIKey()
}
\value{
the saved key
}
\description{
Before calling this function an apiKey needs to be obtained from https://api.data.gov/signup"
and set using setAPIKey
This method loads the saved apiKey from key.rda
}

