% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperAPI.R
\name{getByYear}
\alias{getByYear}
\title{Subsets data by year}
\usage{
getByYear(year, dataDf)
}
\arguments{
\item{year}{Year to subset by}

\item{dataDf}{Data to subset from}
}
\value{
character
}
\description{
Subsets data by year
}

