% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperAPI.R
\name{getDataPerSchool}
\alias{getDataPerSchool}
\title{Helper function to get data based on school name if data for the specified school is available}
\usage{
getDataPerSchool(schoolName, dataDf)
}
\arguments{
\item{schoolName}{Name of school}

\item{dataDf}{Dataframe to get data from}
}
\value{
data.frame
}
\description{
Helper function to get data based on school name if data for the specified school is available
}

