% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperAPI.R
\name{getLastElement}
\alias{getLastElement}
\title{Splits a string and returns the last element}
\usage{
getLastElement(x)
}
\arguments{
\item{x}{String to split}
}
\value{
character
}
\description{
Splits a string and returns the last element
}

