% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repaymentRateByIncome.R
\name{repaymentRateByIncome}
\alias{repaymentRateByIncome}
\title{Plots repayment comparison based on income for specified school(s). Here repayment is the fraction of borrowers successfully repaying their loan
x years after entering repayment. If schoolNames is not specified, data for all schools will be returned}
\usage{
repaymentRateByIncome(apiKey, dataset, year = 2013, schoolNames,
  repaymentYears = 3)
}
\arguments{
\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{year}{Year of college scorecard data}

\item{schoolNames}{The names of schoold to be compared}

\item{repaymentYears}{Years after entering repayment}
}
\value{
ggplot2 object
}
\description{
Plots repayment comparison based on income for specified school(s). Here repayment is the fraction of borrowers successfully repaying their loan
x years after entering repayment. If schoolNames is not specified, data for all schools will be returned
}
\examples{
data(scorecard13)
repaymentRateByIncome(dataset = scorecard13, year = 2013, 
schoolNames = c("University of Massachusetts-Lowell","Massachusetts Institute of Technology", "Drake University"))
}

