% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scorecard13}
\alias{scorecard13}
\title{2013 Scorecard}
\format{A dataframe of 2,133 observations and 1,729 variables}
\source{
\url{https://collegescorecard.ed.gov/data/documentation/}
}
\usage{
scorecard13
}
\description{
2013 College Scorecard data including all available variables and subset to colleges that
predominantly offer an Bachelor's Degree. 
Refer to dataDict for a data dictionary of fields.
}
\examples{
data(scorecard13)
top5Degrees(,scorecard13,c("Stanford University"))

}
\keyword{datasets}

