% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecardInfo.R
\name{scorecardInfo}
\alias{scorecardInfo}
\title{Returns the variable name, definitions, source and any notes from the College Scorecard data dictionary}
\usage{
scorecardInfo(variable)
}
\arguments{
\item{variable}{Provide the variable(s) to return their definition and characteristics}
}
\description{
Returns the variable name, definitions, source and any notes from the College Scorecard data dictionary
}
\examples{
scorecardInfo(c("CITY","INSTNM","UNITID"))
}

