% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentIncomeBy.R
\name{studentIncomeBy}
\alias{studentIncomeBy}
\title{Returns a bar chart of family income based on the student type (aided, dependent, independent).
If using API please provide apiKey and leave dataset field blank. If using a dataset, please leave 
apiKey field blank. Dataset must use the same naming conventions are the College Scorecard.}
\usage{
studentIncomeBy(apiKey, dataset, schools, bygroup, year = "2013")
}
\arguments{
\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}

\item{bygroup}{Populate with aided, dependent or independent to view family income for those student subsets.}

\item{year}{Year of data request. Default is 2013.}
}
\description{
Returns a bar chart of family income based on the student type (aided, dependent, independent).
If using API please provide apiKey and leave dataset field blank. If using a dataset, please leave 
apiKey field blank. Dataset must use the same naming conventions are the College Scorecard.
}
\examples{
data(scorecard13)
studentIncomeBy(,scorecard13,c("Boston University","Northeastern University"),"dependent")
}

