% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completionRate.R, R/debtBoxplot.R, R/medianDebtBy.R, R/studentIncomeBy.R, R/studentRace.R, R/top5Degrees.R
\name{subsetToCategory}
\alias{subsetToCategory}
\title{Returns a bar chart for median debt levels for selected demographic groups by school. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard. Leave bygroup field blank if only interested
overall median.}
\usage{
subsetToCategory(category, apiKey, dataset, schools)

subsetToCategory(category, apiKey, dataset, schools)

subsetToCategory(category, apiKey, dataset, schools)

subsetToCategory(category, apiKey, dataset, schools)

subsetToCategory(category, apiKey, dataset, schools)

subsetToCategory(category, apiKey, dataset, schools)
}
\arguments{
\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}

\item{bygroup}{Leave this blank to see completion rates overall, otherwise populate with 'race' to see completion by race}

\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}

\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}

\item{bygroup}{Leave this blank to see median debt overall, otherwise populate with any of the following bygroups to see
median debt by category: completion, income, gender, Pell, dependence, firstGen}

\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}

\item{bygroup}{Populate with aided, dependent or independent to view family income for those student subsets.}

\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}

\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}
}
\description{
Returns a bar chart for median debt levels for selected demographic groups by school. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard. Leave bygroup field blank if only interested
overall median.

Returns a box plot of the 10th, 25th, 50th, 75th, and 90th percentiles of student debt by school. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard.

Returns a bar chart for median debt levels for selected demographic groups by school. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard. Leave bygroup field blank if only interested
overall median.

Returns a bar chart of family income based on the student type (aided, dependent, independent).
If using API please provide apiKey and leave dataset field blank. If using a dataset, please leave 
apiKey field blank. Dataset must use the same naming conventions are the College Scorecard.

Returns a horizontal stacked bar chart of the racial makeup of the selected schools. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard.

Returns the top five degrees for each school selected and their proportion. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard.
}
\examples{
completionRate(,scorecard13,c("University of Chicago","Northwestern University"),"race")
debtBoxplot(,scorecard13,c("New York University","Cornell University"))
medianDebtBy(,CS2013,c("Hampshire College","Amherst College"),"firstGen")
studentIncomeBy(,CS2013,c("Boston University","Northeastern University"),"dependent")
studentRace(,CS2013,c("Yale University","Harvard University"))
top5 <- top5Degrees(,scorecard13,c("Stanford University"))
}

