% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costByIncome.R
\name{costByIncome}
\alias{costByIncome}
\title{Plots repayment comparison based on income for specified school(s). Here repayment is the fraction of borrowers successfully repaying their loan
x years after entering repayment. If schoolNames is not specified, data for all schools will be returned}
\usage{
costByIncome(apiKey, dataset, year = 2013, schoolNames, type = "public")
}
\arguments{
\item{apiKey}{If using API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{year}{Year of college scorecard data}

\item{schoolNames}{The names of schoold to be compared}

\item{type}{Type of institution. CUrrently the available options are: public, private}
}
\value{
ggplot2 object
}
\description{
Plots repayment comparison based on income for specified school(s). Here repayment is the fraction of borrowers successfully repaying their loan
x years after entering repayment. If schoolNames is not specified, data for all schools will be returned
}
\examples{
data(scorecard13)
costByIncome(dataset = scorecard13, year = 2013, schoolNames = 
c("University of Massachusetts-Lowell", "University of Arizona", "Alabama A & M University"))
costByIncome(dataset = scorecard13, year = 2013, schoolNames = c("Massachusetts Institute of 
Technology","Drake University", "Stanford University"), type = "private")
}

