% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllDataInCategory.R
\name{getAllDataInCategory}
\alias{getAllDataInCategory}
\title{Get all data from a specified category. The available categories are 
root, school, academics, admissions, student, cost, aid, repayment, completion, earnings}
\usage{
getAllDataInCategory(apiKey, dataset, categoryName, year, pattern = "",
  addParams = "id,school.name")
}
\arguments{
\item{apiKey}{Key used for interacting with the API}

\item{dataset}{If not using API please provide dataset name here}

\item{categoryName}{The category of data for retrieval. Options include: academics, admission, aid, cost, earnings, repayment,
root, school, or student.}

\item{year}{The year(s) for data retrieval.}

\item{pattern}{Common text in name of parameters of interest}

\item{addParams}{Parameters outside of category to include in data retrieval}
}
\value{
character vector
}
\description{
Get all data from a specified category. The available categories are 
root, school, academics, admissions, student, cost, aid, repayment, completion, earnings
}
\examples{
\dontrun{costData <- getAllDataInCategory(categoryName = "cost", year = 2013)}
\dontrun{schoolData <- getAllDataInCategory(categoryName = "school")}
\dontrun{earningsData <- getAllDataInCategory(categoryName = "earnings", 
year = c(2010, 2013))}
\dontrun{earningsData <- getAllDataInCategory(categoryName = "earnings", 
  year = c(2010, 2013), pattern = "6_yrs_after_entry.mean", addParams = "school.state")}
}

