% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoAboutSchool.R
\name{getInfoAboutSchool}
\alias{getInfoAboutSchool}
\title{Returns information about a given school}
\usage{
getInfoAboutSchool(schoolName, dataset)
}
\arguments{
\item{schoolName}{Name of school}
}
\description{
Returns information about a given school
}
\examples{
data(scorecard13)
getInfoAboutSchool("University of Massachusetts-Lowell")
}

