% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoAboutSchool.R
\name{getInfoAboutSchool}
\alias{getInfoAboutSchool}
\title{Returns information about a given school}
\usage{
getInfoAboutSchool(apiKey, dataset, schoolName)
}
\arguments{
\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schoolName}{Name of school}
}
\description{
Returns information about a given school
}
\examples{
data(scorecard13)
getInfoAboutSchool(,scorecard13, schoolName = 
c("University of Massachusetts-Lowell", "Stanford University"))
}

