% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetToCategory.R
\name{subsetToCategory}
\alias{subsetToCategory}
\title{Subsets included College Scorecard data to specific category and 
prepares for data visualization.}
\usage{
subsetToCategory(categoryName, apiKey, dataset, schools)
}
\arguments{
\item{categoryName}{The category of data for retrieval. Options include: academics, 
admission, aid, completion, cost, earnings, repayment, root, school, or student.}

\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}
}
\description{
Subsets included College Scorecard data to specific category and 
prepares for data visualization.
}
\examples{
subsetToCategory("aid",,scorecard13,c("Hampshire College","Amherst College"))
}

