% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top5Degrees.R
\name{top5Degrees}
\alias{top5Degrees}
\title{Returns the top five degrees for each school selected and their proportion. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard.}
\usage{
top5Degrees(apiKey, dataset, schools)
}
\arguments{
\item{apiKey}{If use API please provide saved API key value here}

\item{dataset}{If not using API please provide dataset name here}

\item{schools}{The schools you are retrieving data for}
}
\description{
Returns the top five degrees for each school selected and their proportion. If using API please provide
apiKey and leave dataset field blank. If using a dataset, please leave apiKey field blank. Dataset must
use the same naming conventions are the College Scorecard.
}
\examples{
data(scorecard13)
top5 <- top5Degrees(,scorecard13,c("Stanford University"))
}

